REFACTOR: 데이터 초기화 로직을 DevInitializer 클래스로 분리하여 dev 프로필에서만 실행되도록 수정
