+++ src/main/resources/application-dev.yml
@@ -15,9 +15,9 @@ spring:
       ddl-auto: create
     properties:
       hibernate:
+        dialect: org.hibernate.dialect.H2Dialect
         # show_sql: true
         format_sql: true
-        dialect: org.hibernate.dialect.H2Dialect
 
 server:
   port: 8081
